requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.58.4
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyasn==1.6.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scipy==1.13.1
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
zipp==3.23.0
